requirement.txt
asgiref==3.4.1
asttokens==2.0.5
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
black==21.9b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.7.0
Django==4.0.2
django-ckeditor==6.3.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-import-export==2.7.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
executing==0.8.3
factory-boy==3.2.0
Faker==10.0.0
Flask==2.0.3
gunicorn==20.1.0
h11==0.12.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
ipython==8.1.1
itsdangerous==2.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==3.2.0
MarkupPy==1.14
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
paramiko==2.10.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
prettytable==3.2.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.1
pytest==6.2.5
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
PyYAML==5.4.1
regex==2021.11.10
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.2.0
tablib==3.2.0
text-unidecode==1.3
texttable==1.6.4
toml==0.10.2
tomli==1.2.2
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
websocket-client==0.59.0
Werkzeug==2.0.3
whitenoise==6.0.0
wsproto==1.0.0
xlrd==2.0.1
xlwt==1.3.0
